add docs to sdk about panicking in validate contexts for some functions
