+++ fvm/src/kernel/default.rs
@@ -210,6 +210,8 @@ where
     }
 
     fn current_balance(&self) -> Result<TokenAmount> {
+        crate::assert_validator!(self, "Validators don't have a balance.");
+
         // If the actor doesn't exist, it has zero balance.
         Ok(self.get_self()?.map(|a| a.balance).unwrap_or_default())
     }
@@ -295,6 +297,9 @@ where
     }
 
     fn block_create(&mut self, codec: u64, data: &[u8]) -> Result<BlockId> {
+        // TODO may just no-op instead of err here
+        crate::assert_validator!(self, "Validator can't mutate state.");
+
         self.call_manager
             .charge_gas(self.call_manager.price_list().on_block_create(data.len()))?;
 
@@ -302,6 +307,9 @@ where
     }
 
     fn block_link(&mut self, id: BlockId, hash_fun: u64, hash_len: u32) -> Result<Cid> {
+        // TODO may just no-op instead of err here
+        crate::assert_validator!(self, "Validator can't call send.");
+
         if hash_fun != BLAKE2B_256 || hash_len != 32 {
             return Err(syscall_error!(IllegalCid; "cids must be 32-byte blake2b").into());
         }
@@ -453,6 +461,7 @@ where
     C: CallManager,
 {
     fn total_fil_circ_supply(&self) -> Result<TokenAmount> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         // From v15 and onwards, Filecoin mainnet was fixed to use a static circ supply per epoch.
         // The value reported to the FVM from clients is now the static value,
         // the FVM simply reports that value to actors.
@@ -518,6 +527,7 @@ where
         proof_type: RegisteredSealProof,
         pieces: &[PieceInfo],
     ) -> Result<Cid> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager.charge_gas(
             self.call_manager
                 .price_list()
@@ -531,6 +541,7 @@ where
 
     /// Verify seal proof for sectors. This proof verifies that a sector was sealed by the miner.
     fn verify_seal(&mut self, vi: &SealVerifyInfo) -> Result<bool> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .charge_gas(self.call_manager.price_list().on_verify_seal(vi))?;
 
@@ -540,6 +551,7 @@ where
     }
 
     fn verify_post(&mut self, verify_info: &WindowPoStVerifyInfo) -> Result<bool> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .charge_gas(self.call_manager.price_list().on_verify_post(verify_info))?;
 
@@ -553,6 +565,7 @@ where
         h2: &[u8],
         extra: &[u8],
     ) -> Result<Option<ConsensusFault>> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .charge_gas(self.call_manager.price_list().on_verify_consensus_fault())?;
 
@@ -576,6 +589,7 @@ where
     }
 
     fn batch_verify_seals(&mut self, vis: &[SealVerifyInfo]) -> Result<Vec<bool>> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         // NOTE: gas has already been charged by the power actor when the batch verify was enqueued.
         // Lotus charges "virtual" gas here for tracing only.
         log::debug!("batch verify seals start");
@@ -621,6 +635,7 @@ where
         &mut self,
         aggregate: &AggregateSealVerifyProofAndInfos,
     ) -> Result<bool> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager.charge_gas(
             self.call_manager
                 .price_list()
@@ -632,6 +647,7 @@ where
     }
 
     fn verify_replica_update(&mut self, replica: &ReplicaUpdateInfo) -> Result<bool> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager.charge_gas(
             self.call_manager
                 .price_list()
@@ -656,6 +672,7 @@ where
     }
 
     fn charge_gas(&mut self, name: &str, compute: Gas) -> Result<()> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .gas_tracker_mut()
             .charge_gas(name, compute)
@@ -670,7 +687,7 @@ impl<C> NetworkOps for DefaultKernel<C>
 where
     C: CallManager,
 {
-    fn network_epoch(&self) -> ChainEpoch {
+    fn network_epoch(&self) -> ChainEpoch {        
         self.call_manager.context().epoch
     }
 
@@ -680,6 +697,24 @@ where
 
     fn network_base_fee(&self) -> &TokenAmount {
         &self.call_manager.context().base_fee
+    }
+}
+
+impl<C: CallManager> InvokeContextOps for DefaultKernel<C> {
+    fn invoke_context(&self) -> Result<InvocationContext> {
+        Ok(InvocationContext {
+            caller: self.msg_caller(),
+            origin: self.msg_origin().0,
+            receiver: self.msg_receiver(),
+            method_number: self.msg_method_number(),
+            value_received: anyhow::Context::context(
+                self.msg_value_received().try_into(),
+                "invalid token amount",
+            )
+            .or_fatal()?,
+            network_curr_epoch: self.network_epoch(),
+            network_version: self.network_version() as u32,
+        })
     }
 }
 
@@ -693,6 +728,7 @@ where
         rand_epoch: ChainEpoch,
         entropy: &[u8],
     ) -> Result<[u8; RANDOMNESS_LENGTH]> {
+        crate::assert_validator!(self, "Validator can't use randomness.");
         self.call_manager.charge_gas(
             self.call_manager
                 .price_list()
@@ -713,6 +749,8 @@ where
         rand_epoch: ChainEpoch,
         entropy: &[u8],
     ) -> Result<[u8; RANDOMNESS_LENGTH]> {
+        crate::assert_validator!(self, "Validator can't use randomness.");
+
         self.call_manager.charge_gas(
             self.call_manager
                 .price_list()
@@ -733,10 +771,14 @@ where
     C: CallManager,
 {
     fn resolve_address(&self, address: &Address) -> Result<Option<ActorID>> {
+        crate::assert_validator!(self, "Validator can't resolve addresses.");
+
         self.call_manager.state_tree().lookup_id(address)
     }
 
     fn get_actor_code_cid(&self, id: ActorID) -> Result<Option<Cid>> {
+        crate::assert_validator!(self, "Validator can't TODO.");
+
         Ok(self
             .call_manager
             .state_tree()
@@ -748,6 +790,7 @@ where
 
     // TODO(M2) merge new_actor_address and create_actor into a single syscall.
     fn new_actor_address(&mut self) -> Result<Address> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         let origin_addr = *self.call_manager.origin().1;
         let oa = self
             .resolve_to_key_addr(&origin_addr, false)
@@ -769,6 +812,7 @@ where
 
     // TODO(M2) merge new_actor_address and create_actor into a single syscall.
     fn create_actor(&mut self, code_id: Cid, actor_id: ActorID) -> Result<()> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         // TODO https://github.com/filecoin-project/builtin-actors/issues/492
         let singleton = self
             .call_manager
@@ -798,6 +842,7 @@ where
     }
 
     fn get_builtin_actor_type(&self, code_cid: &Cid) -> u32 {
+        // crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .machine()
             .builtin_actors()
@@ -805,6 +850,7 @@ where
     }
 
     fn get_code_cid_for_type(&self, typ: u32) -> Result<Cid> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         self.call_manager
             .machine()
             .builtin_actors()
@@ -816,6 +862,7 @@ where
 
     #[cfg(feature = "m2-native")]
     fn install_actor(&mut self, code_id: Cid) -> Result<()> {
+        crate::assert_validator!(self, "Validator can't TODO.");
         // TODO figure out gas
         self.call_manager
             .machine()
