+++ fvm/src/syscalls/network.rs
@@ -6,6 +6,8 @@ use crate::kernel::{ClassifyResult, Kernel, Result};
 
 /// Returns the base fee split as two u64 ordered in little endian.
 pub fn base_fee(context: Context<'_, impl Kernel>) -> Result<sys::TokenAmount> {
+    crate::assert_validator!(context.kernel, "Validator can't TODO.");
+
     context
         .kernel
         .network_base_fee()
@@ -16,6 +18,8 @@ pub fn base_fee(context: Context<'_, impl Kernel>) -> Result<sys::TokenAmount> {
 
 /// Returns the network circ supply split as two u64 ordered in little endian.
 pub fn total_fil_circ_supply(context: Context<'_, impl Kernel>) -> Result<sys::TokenAmount> {
+    crate::assert_validator!(context.kernel, "Validator can't TODO.");
+
     context
         .kernel
         .total_fil_circ_supply()?
