+++ sdk/src/network.rs
@@ -7,10 +7,14 @@ use fvm_shared::version::NetworkVersion;
 use crate::sys;
 use crate::vm::INVOCATION_CONTEXT;
 
+
+/// Panics inside validate context
 pub fn curr_epoch() -> ChainEpoch {
     INVOCATION_CONTEXT.network_curr_epoch
 }
 
+
+/// Panics inside validate context
 pub fn version() -> NetworkVersion {
     INVOCATION_CONTEXT
         .network_version
@@ -18,6 +22,7 @@ pub fn version() -> NetworkVersion {
         .expect("invalid network version")
 }
 
+/// Panics inside validate context
 pub fn base_fee() -> TokenAmount {
     unsafe {
         sys::network::base_fee()
@@ -26,6 +31,7 @@ pub fn base_fee() -> TokenAmount {
     }
 }
 
+/// Panics inside validate context
 pub fn total_fil_circ_supply() -> TokenAmount {
     unsafe {
         sys::network::total_fil_circ_supply()
