+++ sdk/src/sself.rs
@@ -41,6 +41,8 @@ pub fn set_root(cid: &Cid) -> Result<(), NoStateError> {
 }
 
 /// Gets the current balance for the calling actor.
+/// 
+/// Panics inside of Validate contexts.
 #[inline(always)]
 pub fn current_balance() -> TokenAmount {
     unsafe {
